backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
ffmpeg-python==0.2.0
google-generativeai==0.7.2
python-dotenv==1.0.1
aiofiles==23.2.1
